[soundemittersystem] SetChannel & SetDelayMsec should use uint16 as expected

+++ public/SoundEmitterSystem/isoundemittersystembase.h
@@ -147,11 +147,11 @@ struct CSoundParametersInternal
 	bool		UsesGenderToken() const						{ return uses_gender_token; }
 	bool		ShouldPreload() const						{ return m_bShouldPreload; }
 
-	void		SetChannel( int newChannel )				{ channel = newChannel; }
+	void		SetChannel( uint16 newChannel )				{ channel = newChannel; }
 	void		SetVolume( float start, float range = 0.0 )	{ volume.start = start; volume.range = range; }
 	void		SetPitch( uint8 start, uint8 range = 0 )	{ pitch.start = start; pitch.range = range; }
 	void		SetSoundLevel( uint16 start, uint16 range = 0 )	{ soundlevel.start = start; soundlevel.range = range; }
-	void		SetDelayMsec( int delay )					{ delay_msec = delay; }
+	void		SetDelayMsec( uint16 delay )					{ delay_msec = delay; }
 	void		SetShouldPreload( bool bShouldPreload )		{ m_bShouldPreload = bShouldPreload;	}
 	void		SetOnlyPlayToOwner( bool b )				{ play_to_owner_only = b; }
 	void		SetHadMissingWaveFiles( bool b )			{ had_missing_wave_files = b; }
